fix: move completion regression test to its own sibling under the size guard
The new guard test pushed respawn_guard_tests.rs over the Server Guards file
size cap (1500 lines / 51200 bytes); split it into a #[path] sibling like
respawn_guard_mergequeue_tests.rs.
